requirements.txt
aiogram==3.4
feedparser
httpx
asyncpg
apscheduler

googletrans==4.0.0-rc1
google-generativeai